#25 refactor: extract SQL generation into dialect-specific modules
- Create entity/sql/ module structure (Rust 2018+ style, no mod.rs)
- Move PostgreSQL implementation to sql/postgres.rs
- Add sql/clickhouse.rs placeholder with compile_error
- Add sql/mongodb.rs placeholder with compile_error
- Update sql.rs as coordinator that delegates to dialects
- Update expected error messages in fail tests
